WIP: parquet writer updates
